+++ core/framework.js
@@ -8,11 +8,9 @@
 //            http://github.com/mwaylabs/The-M-Project/blob/master/GPL-LICENSE
 // ==========================================================================
 
-var E = require('./e').E,
-Framework,
-File = require('./file').File;
-
-
+var E = require('./e').E;
+var Framework;
+var File = require('./file').File;
 
 /**
  * @class
@@ -41,6 +39,7 @@ Framework = exports.Framework = function (properties) {
   this.app     = null;
   this.virtual = false;
   this.library = false;
+  this.is_a_plugin  = false;
   this.path = '';
   this.name = '';
   this.frDelimiter = '';
@@ -91,44 +90,84 @@ Framework.prototype.isVirtual = function () {
  * Reads the content of the files connected to a framework.
  * Needs a array with File objects, to read the content.
  *
- * @param arrayOfFiles {object}, the Array containing the files
  * @param callback {function}, the callback function.
  */
-Framework.prototype.readFiles = function (arrayOfFiles, callback) {
+Framework.prototype.readFiles = function (callback) {
   var self = this;
+  var resourceCounter = self.files.length;
 
-  if (Array.isArray(arrayOfFiles)) {
-
-    if (arrayOfFiles.length === 0) {
-      callback(null, arrayOfFiles);
-    }
-
-    var _FileReader = function (cb, array) {
-      var that = this;
-      this._resourceCounter = array.length;
-      that.callbackIfDone = function () {
-        if (that._resourceCounter <= 0) {
-          cb(null, self.files);
-        }
-      };
-
-      that.read = function () {
-        array.forEach(function (file) {
-            var _cFPath  = file.path;
-            self._e_.fs.readFile(_cFPath, function (err, data) { // data is a buffer object, if no encoding was specified!
-                if (err) {
-                  throw err;
-                } else {
-                  file.content = data;
-                  that._resourceCounter -= 1;
-                  that.callbackIfDone();
-                }
-              });
-          });
-      };
-    }
-    new _FileReader(callback, arrayOfFiles).read();
+  if (self.files.length === 0) {
+    callback(null, self.files);
   }
+
+  function callbackIfDone() {
+    if (resourceCounter <= 0) {
+      callback(null, self.files);
+    }
+  }
+
+  self.files.forEach(function (file) {
+      var _cFPath  = file.path;
+      self._e_.fs.readFile(_cFPath, function (err, data) { // data is a buffer object, if no encoding was specified!
+          if (err) {
+            throw err;
+          } else {
+            file.content = data;
+            resourceCounter -= 1;
+            callbackIfDone();
+          }
+        });
+    });
+};
+
+/**
+ * @description
+ * Delegates reading of needed files to the appropriate method
+ * @param path, the path to look for resources.
+ * @param callback, the function, that is called after all resources haven been loaded.
+ */
+Framework.prototype.getFiles = function getFiles(path, callback) {
+  var that = this;
+  var manifest = path + '/manifest.json';
+
+  this._e_.fs.stat(manifest, function (err, stat) {
+      if (err) {
+        that.browseFiles(path, callback);
+      } else {
+        that.readManifest(manifest, path, callback);
+      }
+    });
+};
+
+Framework.prototype.readManifest = function readManifest(manifest, path, callback) {
+  var that = this;
+  this._e_.fs.readFile(manifest, 'utf8', function (err, content) {
+      if (err) {
+        throw err;
+      }
+      try {
+        that.files = JSON.parse(content).manifest.map(function (file) {
+            if (that.touchPath(path + '/' + file)) {
+                return new File({
+                    frDelimiter: that.frDelimiter,
+                    name: path + '/' + file,
+                    path: path + '/' + file,
+                    framework: that
+                });
+            } else {
+               console.log("\n");
+               console.log(that.style.red('ERROR:')+that.style.green(' File "')+that.style.cyan(path.split(that.frDelimiter)[1] + '/' + file)
+                          + that.style.green('" was referencd in "') + that.style.cyan(path.split(that.frDelimiter)[1] + '/'+'manifest.json')
+                          + that.style.green('" but not found in directory. '));
+               console.log("\n");
+               process.exit(1); /* exit the process, reason: file not found*/
+            }
+        });
+        callback(null);
+      } catch (ex) {
+        callback(ex);
+      }
+    });
 };
 
 /**
@@ -148,8 +187,8 @@ Framework.prototype.browseFiles = function (path, callback) {
 
     that.callbackIfDone = function () {
       if (that._folderCounter <= 0) {
-        //           console.log('framework '+framework.name+'  _folderCounter '+that._folderCounter);
-        callback(null, framework.files);
+        callback(null);
+
       }
     };
 
@@ -162,11 +201,7 @@ Framework.prototype.browseFiles = function (path, callback) {
       if (that.checkIfFolderShouldBeExcluded(path)) {
         return true;
       }
-      if (self.excludedFiles.indexOf(_fileBaseName[_fileBaseName.length - 1]) === -1) {
-        return false;
-      } else {
-        return true;
-      }
+      return !(self.excludedFiles.indexOf(_fileBaseName[_fileBaseName.length - 1]) === -1);
     };
 
     /*
@@ -207,11 +242,11 @@ Framework.prototype.browseFiles = function (path, callback) {
               if (!that.checkIfFileShouldBeExcluded(path)) {
                 framework.files.push(
                   new File({
-                    frDelimiter: framework.frDelimiter,
-                    name: path,
-                    path: path,
-                    framework: framework
-                  })
+                      frDelimiter: framework.frDelimiter,
+                      name: path,
+                      path: path,
+                      framework: framework
+                    })
                 );
               }
               that.callbackIfDone();
@@ -241,14 +276,17 @@ Framework.prototype.build = function (callback) {
   } else {
     that.sequencer(
       function () {
-        that.browseFiles(that.path, this);
+        that.getFiles(that.path, this);
       },
-      function (err, files) {
+
+      function (err) {
         if (err) {
           throw err;
         }
-        that.readFiles(files, this);
+      //  console.dir(this);
+        that.readFiles(this);
       },
+
       function (err, files) {
         if (err) {
           throw err;
@@ -316,13 +354,15 @@ Framework.prototype.save = function (callback) {
           that.copyFile(files, _cF.path, _outputPath + '/theme/images/' + _cF.getBaseName() + _cF.getFileExtension());
           break;
         case (_cF.isStylesheet()):
+          (_cF.containsMergedContent) ?
+          that.writeFile(files,  _outputPath + '/theme/' + _cF.getBaseName() + _cF.getFileExtension(), _cF.content) :
           that.copyFile(files, _cF.path, _outputPath + '/theme/' + _cF.getBaseName() + _cF.getFileExtension());
           break;
-          /*
-           case (_cF.isSASS_Stylesheet()):
+        /*
+        case (_cF.isSASS_Stylesheet()):
              that.writeFile(files, _outputPath+'/theme/'+_cF.getBaseName()+'.css', _cF.content);
              break; */
-        // case (_cF.isVirtual()):
+     // case (_cF.isVirtual()):
         default:
           var _fileName =  (self.combinedScripts) ? self.name + '.js' : _cF.getBaseName() + _cF.getFileExtension();
           that.writeFile(files, _outputPath + '/' + _fileName, _cF.content);
@@ -342,11 +382,12 @@ Framework.prototype.save = function (callback) {
  * @param callback, the function, that is executed after the prepareForServer() is done.
  */
 Framework.prototype.prepareForServer = function (server, callback) {
+  var appName = this.app.name;
   this.files.forEach(function (file) {
       if (!server.files) {
         server.files = {};
       }
-      server.files['/' + file.requestPath] = file;
+      server.files['/'+appName+'/' + file.requestPath] = file;
     });
   callback();
 };