+++ tasks/task_contentType.js
@@ -40,7 +40,7 @@ Task_ContentType = exports.Task_ContentType = function() {
 };
 
 
-Task_ContentType.prototype = new Task;
+Task_ContentType.prototype = new Task();
 
 
 /**
@@ -50,6 +50,7 @@ Task_ContentType.prototype = new Task;
  */
 Task_ContentType.prototype.duty = function(framework,callback){
 var self = this;
+    
   framework.files.forEach(function(cF){
   cF.contentType = (self.contentTypes[cF.getFileExtension()]) ? self.contentTypes[cF.getFileExtension()] :  'text/plain';